refactor(core): stop_reason_precedence returns StopPrecedence (S-7, #508)
The function computed the private Ord enum and then erased it to u8.
Make StopPrecedence public and return it, so both lattice consumers
(ScopeLifecycle::begin_drain and ScopeCell's stopped publisher) compare
the ordered type directly. Only the exploration walk's packed
fingerprint still wants a small integer, and it casts there. No facade
edit: the facade's comparisons are type-agnostic.

+++ crates/shelterwood-core/src/engine.rs
@@ -844,7 +844,7 @@ impl ScopeLifecycle {
                     StartupPhase::Complete => 1,
                     StartupPhase::Failed => 2,
                 },
-                stop_reason_precedence(reason),
+                stop_reason_precedence(reason) as u8,
             ),
         }
     }
